pyproject.toml
[project]
name = "wreporter"
version = "0.1.0"
description = "B2B 영업팀용 기업 분석 리포트 자동 생성 시스템"
requires-python = ">=3.10"
dependencies = [
    "reflex>=0.6.0",
    "anthropic>=0.40.0",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "pytest>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["wreporter*", "core*", "clients*", "db*", "prompts*", "utils*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
